Revert "Merge pull request #168 from canjs/revert-165-can-observe"
This reverts commit 68c790660645aa3e1e7fd25c001de25e33e80cc7, reversing
changes made to 5aca58a7aa3d74bd706955806f5135e27d234480.
