+++ .npmignore
@@ -1,1 +1,2 @@
 !dist/
+!test/
