--- build.js
@@ -1,21 +1,0 @@
-var stealTools = require("steal-tools");
-
-stealTools.export({
-	steal: {
-		config: __dirname + "/package.json!npm"
-	},
-	outputs: {
-		"+amd": {},
-		"+standalone": {
-			exports: {
-				"can-namespace": "can"
-			}
-		}
-	}
-}).catch(function(e){
-	
-	setTimeout(function(){
-		throw e;
-	},1);
-	
-});
