+++ control/control.js
@@ -1,9 +1,6 @@
 var Control = require("can-control");
-
 var canEach = require('can-util/js/each/each');
-var string = require('can-util/js/string/string');
-var canCompute = require("can-compute");
-var observeReader = require('can-stache-key');
+var canReflect = require("can-reflect");
 
 // ## Helpers
 // Attribute names to ignore for setting viewModel values.
@@ -56,7 +53,7 @@ var ComponentControl = Control.extend({
 			// unbind `change` event listeners set by the controller.
 			if (this._bindings) {
 				canEach(this._bindings.readyComputes || {}, function(value) {
-					value.compute.unbind("change", value.handler);
+					canReflect.offValue(value.compute, value.handler);
 				});
 			}
 			// Call `Control.prototype.off` function on this instance to cleanup the bindings.
