+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "can-component",
-  "version": "3.3.6",
+  "version": "4.0.0-pre.6",
   "description": "Custom elements for CanJS",
   "homepage": "http://canjs.com",
   "repository": {
@@ -13,9 +13,8 @@
     "url": "http://bitovi.com"
   },
   "scripts": {
-    "preversion": "npm test && npm run build",
-    "version": "git commit -am \"Update dist for release\" && git checkout -b release && git add -f dist/",
-    "postversion": "git push --tags && git checkout master && git branch -D release && git push",
+    "preversion": "npm test",
+    "postpublish": "git push --tags && git push",
     "testee": "testee test/test.html --browsers firefox",
     "test": "npm run detect-cycle && npm run jshint && npm run testee",
     "jshint": "jshint ./*.js --config",
@@ -23,7 +22,6 @@
     "release:patch": "npm version patch && npm publish",
     "release:minor": "npm version minor && npm publish",
     "release:major": "npm version major && npm publish",
-    "build": "node build.js",
     "develop": "done-serve --static --develop --port 8080",
     "detect-cycle": "detect-cyclic-packages --ignore done-serve"
   },
@@ -39,29 +37,28 @@
     ]
   },
   "dependencies": {
-    "can-compute": "^3.3.1",
     "can-construct": "^3.2.0",
-    "can-control": "^3.2.0",
-    "can-globals": "^0.2.3",
+    "can-control": "^4.0.0-pre.2",
+    "can-globals": "<2.0.0",
     "can-namespace": "1.0.0",
-    "can-observation": "^3.3.1",
-    "can-reflect": "^1.2.1",
-    "can-simple-map": "^3.3.0",
-    "can-stache-bindings": "^3.4.0",
-    "can-stache-key": "0.1.0",
-    "can-types": "^1.1.0",
+    "can-observation": "^4.0.0-pre.11",
+    "can-observe": "^2.0.0-pre.0",
+    "can-queues": "<2.0.0",
+    "can-reflect": "^1.6.0",
+    "can-simple-map": "^4.0.0-pre.0",
+    "can-simple-observable": "^2.0.0-pre.7",
+    "can-stache-bindings": "^4.0.0-pre.7",
+    "can-stache-key": "^1.0.0-pre.2",
     "can-util": "^3.9.5",
-    "can-view-callbacks": "^3.2.0",
-    "can-view-model": "^3.4.0",
+    "can-view-callbacks": "^4.0.0-pre.0",
+    "can-view-model": "^4.0.0-pre.1",
     "can-view-nodelist": "^3.1.0",
-    "can-view-scope": "^3.3.1"
+    "can-view-scope": "^4.0.0-pre.16"
   },
   "devDependencies": {
-    "can-define": "^1.3.0",
-    "can-event": "^3.5.0",
-    "can-list": "^3.2.0",
-    "can-map": "^3.3.1",
-    "can-stache": "^3.3.0",
+    "can-define": "^2.0.0-pre.1",
+    "can-stache": "^4.0.0-pre.2",
+    "can-symbol": "^1.4.1",
     "can-vdom": "^3.1.0",
     "detect-cyclic-packages": "^1.1.0",
     "done-serve": "^1.3.0",
