+++ test/component-helpers-test.js
@@ -1,0 +1,34 @@
+var QUnit = require("steal-qunit");
+
+var helpers = require("./helpers");
+var SimpleMap = require("can-simple-map");
+var stache = require("can-stache");
+var Component = require("can-component");
+var domEvents = require('can-util/dom/events/events');
+
+var innerHTML = function(el){
+    return el && el.innerHTML;
+};
+
+helpers.makeTests("can-component helpers", function(){
+
+    QUnit.test("helpers reference the correct instance (#515)", function () {
+		expect(2);
+		Component({
+			tag: 'my-text',
+			view: stache('<p>{{valueHelper}}</p>'),
+			helpers: {
+				valueHelper: function () {
+					return this.value;
+				}
+			}
+		});
+
+		var renderer = stache('<my-text value:from="\'value1\'"></my-text><my-text value:from="\'value2\'"></my-text>');
+
+		var frag = renderer({});
+
+		equal(frag.firstChild.firstChild.firstChild.nodeValue, 'value1');
+		equal(frag.lastChild.firstChild.firstChild.nodeValue, 'value2');
+	});
+});
