--- test/component-test.js
@@ -1,3 +1,0 @@
-require('./component-define-test');
-require('./component-map-test');
-require('./component-slot-test');
