+++ test/test.html
@@ -1,4 +1,4 @@
 <!DOCTYPE html>
 <title>can-component</title>
-<script src="../node_modules/steal/steal.js" main="can-component/test/component-test"></script>
+<script src="../node_modules/steal/steal.js" main="can-component/test/tests"></script>
 <div id="qunit-fixture"></div>
