+++ test/tests.js
@@ -1,0 +1,10 @@
+require("./component-tag-test");
+require("./component-viewmodel-test");
+require("./component-viewmodel-observe-test");
+require("./component-view-test");
+require("./component-helpers-test");
+require("./component-events-test");
+require("./example-test");
+require('./component-slot-test');
+require('./component-automount-test');
+require('./component-define-test');
